requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
ansicolors==1.1.8         # via -r requirements.in
certifi==2020.11.8        # via requests
chardet==3.0.4            # via requests
colorama==0.4.4           # via crayons
configparser==5.0.1       # via webdriver-manager
crayons==0.4.0            # via webdriver-manager
idna==2.10                # via requests
requests==2.25.0          # via webdriver-manager
selenium==3.141.0         # via -r requirements.in
tqdm==4.54.1              # via -r requirements.in
urllib3==1.26.2           # via requests, selenium
webdriver-manager==3.2.2  # via -r requirements.in
